+++ src/Writer/UserWriter.php
@@ -52,9 +52,12 @@ class UserWriter implements WriterInterface
     /**
      * @throws Throwable
      */
-    public function save(ObjectInterface $object, PromiseInterface $promise = null): WriterInterface
-    {
-        $this->persist($object, $promise);
+    public function save(
+        ObjectInterface $object,
+        PromiseInterface $promise = null,
+        ?bool $prefereRealDateOnUpdate = null,
+    ): WriterInterface {
+        $this->persist($object, $promise, $prefereRealDateOnUpdate);
 
         return $this;
     }